% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{acs12_tbl_df}
\alias{acs12_tbl_df}
\title{American Community Survey 2012}
\format{
A tibble with 2,000 observations and 13 variables:
\describe{
  \item{income}{Income of individuals (integer).}
  \item{employment}{Employment status (factor with 3 levels).}
  \item{hrs_work}{Number of hours worked per week (integer).}
  \item{race}{Race of individuals (factor with 4 levels).}
  \item{age}{Age of individuals (integer).}
  \item{gender}{Gender of individuals (factor with 2 levels: "male", "female").}
  \item{citizen}{Citizenship status (factor with 2 levels: "no", "yes").}
  \item{time_to_work}{Time taken to travel to work in minutes (integer).}
  \item{lang}{Primary language spoken at home (factor with 2 levels: "english", "other").}
  \item{married}{Marital status (factor with 2 levels: "no", "yes").}
  \item{edu}{Educational attainment (factor with 3 levels).}
  \item{disability}{Disability status (factor with 2 levels).}
  \item{birth_qrtr}{Birth quarter of individuals (factor with 4 levels).}
}
}
\source{
American Community Survey, 2012.
}
\usage{
data(acs12_tbl_df)
}
\description{
The dataset name has been changed to 'acs12_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
'usdatasets' package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
identifies the dataset as a tibble data frame, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
