% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{age_at_mar_tbl_df}
\alias{age_at_mar_tbl_df}
\title{Age at first marriage of 5,534 US women.}
\format{
A tibble with 5,534 observations and 1 variable:
\describe{
  \item{age}{Age at first marriage (integer).}
}
}
\source{
United States Census Data.
}
\usage{
data(age_at_mar_tbl_df)
}
\description{
The dataset name has been changed to 'age_at_mar_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
'usdatasets' package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
identifies the dataset as a tibble data frame, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
