% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{airlines_tbl_df}
\alias{airlines_tbl_df}
\title{Airline names - U.S. Airlines Carrier Codes and Names}
\format{
A tibble with 16 observations and 2 variables:
\describe{
  \item{carrier}{Carrier code (character) representing the airline.}
  \item{name}{Name of the airline (character).}
}
}
\source{
U.S. Department of Transportation.
}
\usage{
data(airlines_tbl_df)
}
\description{
The dataset name has been changed to 'airlines_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
'usdatasets' package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
identifies the dataset as a tibble, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
