% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{airports_tbl_df}
\alias{airports_tbl_df}
\title{Airport metadata - U.S. Airports Information}
\format{
A tibble with 1,458 observations and 8 variables:
\describe{
  \item{faa}{FAA airport code (character).}
  \item{name}{Name of the airport (character).}
  \item{lat}{Latitude of the airport (numeric).}
  \item{lon}{Longitude of the airport (numeric).}
  \item{alt}{Altitude of the airport (numeric).}
  \item{tz}{Time zone (numeric).}
  \item{dst}{Daylight saving time flag (character).}
  \item{tzone}{Time zone name (character).}
}
}
\source{
U.S. Federal Aviation Administration (FAA).
}
\usage{
data(airports_tbl_df)
}
\description{
The dataset name has been changed to 'airports_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
'usdatasets' package and assists users in identifying its specific characteristics. The suffix 'tbl_df'
identifies the dataset as a tibble, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
