% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{airquality_df}
\alias{airquality_df}
\title{New York Air Quality Measurements}
\format{
A data frame with 153 observations and 6 variables:
\describe{
  \item{Ozone}{Ozone concentration (parts per billion) from 1 to 331.}
  \item{Solar.R}{Solar radiation (watts per square meter).}
  \item{Wind}{Wind speed (miles per hour).}
  \item{Temp}{Temperature (degrees Fahrenheit).}
  \item{Month}{Month of the observation (integer from 5 to 9).}
  \item{Day}{Day of the observation (integer from 1 to 31).}
}
}
\source{
United States Environmental Protection Agency (EPA).
}
\usage{
data(airquality_df)
}
\description{
The dataset name has been changed to 'airquality_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
'usdatasets' package and assists users in identifying its specific characteristics. The suffix 'df'
identifies the dataset as a data frame, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
