% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{births14_tbl_df}
\alias{births14_tbl_df}
\title{US Births 2014}
\format{
A tibble with 1,000 observations and 13 variables:
\describe{
  \item{fage}{Age of the father (in years).}
  \item{mage}{Age of the mother (in years).}
  \item{mature}{Indicates if the mother is mature (yes/no).}
  \item{weeks}{Number of weeks of pregnancy.}
  \item{premie}{Indicates if the baby is a premature birth (yes/no).}
  \item{visits}{Number of prenatal visits.}
  \item{gained}{Weight gained by the mother during pregnancy (in pounds).}
  \item{weight}{Birth weight of the baby (in grams).}
  \item{lowbirthweight}{Indicates if the baby is of low birth weight (yes/no).}
  \item{sex}{Sex of the baby (male/female).}
  \item{habit}{Maternal smoking habits (yes/no).}
  \item{marital}{Marital status of the mother (married/single).}
  \item{whitemom}{Indicates if the mother is white (yes/no).}
}
}
\source{
National Vital Statistics Reports
}
\usage{
data(births14_tbl_df)
}
\description{
The dataset name has been changed to 'births14_tbl_df' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original content of the dataset has not been modified in any way.
}
