% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{births_tbl_df}
\alias{births_tbl_df}
\title{North Carolina births, 100 cases}
\format{
A tibble with 150 observations and 9 variables:
\describe{
  \item{f_age}{Age of the father (in years).}
  \item{m_age}{Age of the mother (in years).}
  \item{weeks}{Number of weeks of pregnancy.}
  \item{premature}{Indicates if the baby is premature (factor: yes/no).}
  \item{visits}{Number of prenatal visits.}
  \item{gained}{Weight gained by the mother during pregnancy (in pounds).}
  \item{weight}{Birth weight of the baby (in grams).}
  \item{sex_baby}{Sex of the baby (factor: male/female).}
  \item{smoke}{Indicates if the mother smoked during pregnancy (factor: yes/no).}
}
}
\source{
National Vital Statistics Reports
}
\usage{
data(births_tbl_df)
}
\description{
The dataset name has been changed to 'births_tbl_df' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original content of the dataset has not been modified in any way.
}
