% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Boston_df}
\alias{Boston_df}
\title{Housing Values in Suburbs of Boston}
\format{
A data frame with 506 observations and 14 variables:
\describe{
  \item{crim}{Per capita crime rate by town.}
  \item{zn}{Proportion of residential land zoned for lots over 25,000 sq. ft.}
  \item{indus}{Proportion of non-retail business acres per town.}
  \item{chas}{Charles River dummy variable (1 if tract bounds river; 0 otherwise).}
  \item{nox}{Nitric oxides concentration (parts per 10 million).}
  \item{rm}{Average number of rooms per dwelling.}
  \item{age}{Proportion of owner-occupied units built prior to 1940.}
  \item{dis}{Weighted distances to five Boston employment centers.}
  \item{rad}{Index of accessibility to radial highways.}
  \item{tax}{Full-value property tax rate per $10,000.}
  \item{ptratio}{Pupil-teacher ratio by town.}
  \item{black}{1000(Bk - 0.63)^2 where Bk is the proportion of Black residents by town.}
  \item{lstat}{Percentage of lower status of the population.}
  \item{medv}{Median value of owner-occupied homes in $1000s.}
}
}
\source{
Boston Housing Data
}
\usage{
data(Boston_df)
}
\description{
The dataset name has been changed to 'Boston_df' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix '_df' identifies the dataset as a data frame. The original content of the dataset has not been modified in any way.
}
