% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Cars93_df}
\alias{Cars93_df}
\title{Data from 93 Cars on Sale in the USA in 1993}
\format{
A data frame with 54 observations and 6 variables:
\describe{
  \item{type}{Type of the car (factor with 3 levels).}
  \item{price}{Price of the car (in US dollars).}
  \item{mpg_city}{Miles per gallon in the city.}
  \item{drive_train}{Drive train type (factor with 3 levels).}
  \item{passengers}{Number of passengers the car can accommodate.}
  \item{weight}{Weight of the car (in pounds).}
}
}
\source{
1993 Cars Data
}
\usage{
data(Cars93_df)
}
\description{
The dataset name has been changed to 'Cars93_df' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix '_df' identifies the dataset as a data frame. The original content of the dataset has not been modified in any way.
}
