% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{cia_factbook_tbl_df}
\alias{cia_factbook_tbl_df}
\title{CIA Factbook Details on Countries}
\format{
A tibble with 259 observations and 11 variables:
\describe{
  \item{country}{Name of the country (factor with 259 levels).}
  \item{area}{Total area of the country (in square kilometers).}
  \item{birth_rate}{Birth rate (number of live births per 1,000 people).}
  \item{death_rate}{Death rate (number of deaths per 1,000 people).}
  \item{infant_mortality_rate}{Infant mortality rate (number of deaths of infants under one year old per 1,000 live births).}
  \item{internet_users}{Number of internet users (in millions).}
  \item{life_exp_at_birth}{Life expectancy at birth (in years).}
  \item{maternal_mortality_rate}{Maternal mortality rate (number of maternal deaths per 100,000 live births).}
  \item{net_migration_rate}{Net migration rate (number of migrants per 1,000 people).}
  \item{population}{Total population of the country.}
  \item{population_growth_rate}{Population growth rate (percentage).}
}
}
\source{
CIA World Factbook
}
\usage{
data(cia_factbook_tbl_df)
}
\description{
The dataset name has been changed to 'cia_factbook_tbl_df' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original content of the dataset has not been modified in any way.
}
