% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{cle_sac_tbl_df}
\alias{cle_sac_tbl_df}
\title{Cleveland and Sacramento Demographic and Income Data (2000)}
\format{
A tibble with 500 observations and 8 variables:
\describe{
  \item{year}{Year of the observation (integer).}
  \item{state}{State of the observation (factor with 2 levels).}
  \item{city}{City of the observation (character).}
  \item{age}{Age of the individual (integer).}
  \item{sex}{Sex of the individual (factor with 2 levels).}
  \item{race}{Race of the individual (character).}
  \item{marital_status}{Marital status of the individual (character).}
  \item{personal_income}{Personal income of the individual (integer).}
}
}
\source{
Cleveland Study
}
\usage{
data(cle_sac_tbl_df)
}
\description{
The dataset name has been changed to 'cle_sac_tbl_df' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original content of the dataset has not been modified in any way.
}
