% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{env_regulation_tbl_df}
\alias{env_regulation_tbl_df}
\title{American Adults on Regulation and Renewable Energy}
\format{
A tibble with 705 observations and 1 variable:
\describe{
  \item{statement}{Environmental regulation statement (character).}
}
}
\source{
Environmental Regulation Study
}
\usage{
data(env_regulation_tbl_df)
}
\description{
The dataset name has been changed to 'env_regulation_tbl_df' to avoid confusion with other
packages in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original
content of the dataset has not been modified in any way.
}
