% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{fcid_tbl_df}
\alias{fcid_tbl_df}
\title{Summary of male heights from USDA Food Commodity Intake Database}
\format{
A tibble with 100 observations and 2 variables:
\describe{
  \item{height}{Height of the individual (numeric).}
  \item{num_of_adults}{Number of adults in the household (integer).}
}
}
\source{
Family Characteristics and Income Study
}
\usage{
data(fcid_tbl_df)
}
\description{
The dataset name has been changed to 'fcid_tbl_df' to avoid confusion with other
packages in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original
content of the dataset has not been modified in any way.
}
