% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{govrace10_tbl_df}
\alias{govrace10_tbl_df}
\title{Election results for 2010 Governor races in the U.S.}
\format{
A tibble with 37 observations and 23 variables:
\describe{
  \item{id}{Identification number (numeric).}
  \item{state}{State name (character).}
  \item{abbr}{State abbreviation (character).}
  \item{name1}{Name of the first candidate (character).}
  \item{perc1}{Percentage of votes for the first candidate (numeric).}
  \item{party1}{Political party of the first candidate (character).}
  \item{votes1}{Number of votes for the first candidate (numeric).}
  \item{name2}{Name of the second candidate (character).}
  \item{perc2}{Percentage of votes for the second candidate (numeric).}
  \item{party2}{Political party of the second candidate (character).}
  \item{votes2}{Number of votes for the second candidate (numeric).}
  \item{name3}{Name of the third candidate (character).}
  \item{perc3}{Percentage of votes for the third candidate (numeric).}
  \item{party3}{Political party of the third candidate (character).}
  \item{votes3}{Number of votes for the third candidate (numeric).}
  \item{name4}{Name of the fourth candidate (character).}
  \item{perc4}{Percentage of votes for the fourth candidate (numeric).}
  \item{party4}{Political party of the fourth candidate (character).}
  \item{votes4}{Number of votes for the fourth candidate (numeric).}
  \item{name5}{Name of the fifth candidate (character).}
  \item{perc5}{Percentage of votes for the fifth candidate (numeric).}
  \item{party5}{Political party of the fifth candidate (character).}
  \item{votes5}{Number of votes for the fifth candidate (numeric).}
}
}
\source{
2010 Gubernatorial Races
}
\usage{
data(govrace10_tbl_df)
}
\description{
The dataset name has been changed to 'govrace10_tbl_df' to avoid confusion with other
packages in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original
content of the dataset has not been modified in any way.
}
