% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{homicides15_tbl_df}
\alias{homicides15_tbl_df}
\title{Homicides in nine cities in 2015}
\format{
A tibble with 1922 observations and 15 variables:
\describe{
  \item{uid}{Unique identifier (integer).}
  \item{city_name}{City name where the homicide occurred (character).}
  \item{offense_code}{Offense code (character).}
  \item{offense_type}{Type of offense (character).}
  \item{date_single}{Date of the homicide (POSIXct).}
  \item{address}{Location address of the homicide (character).}
  \item{longitude}{Longitude of the homicide location (numeric).}
  \item{latitude}{Latitude of the homicide location (numeric).}
  \item{location_type}{Type of location where the homicide occurred (character).}
  \item{location_category}{Category of the location (character).}
  \item{fips_state}{FIPS code of the state (integer).}
  \item{fips_county}{FIPS code of the county (character).}
  \item{tract}{Census tract where the homicide occurred (character).}
  \item{block_group}{Block group number (integer).}
  \item{block}{Block number (integer).}
}
}
\source{
2015 Homicides Data
}
\usage{
data(homicides15_tbl_df)
}
\description{
The dataset name has been changed to 'homicides15_tbl_df' to avoid confusion with other
packages in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package and assists users in identifying
its specific characteristics. The suffix 'tbl_df' identifies the dataset as a tibble. The original
content of the dataset has not been modified in any way.
}
