% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{houserace10_tbl_df}
\alias{houserace10_tbl_df}
\title{Election results for the 2010 U.S. House of Represenatives races}
\format{
A tibble with 435 observations and 24 variables:
\describe{
  \item{id}{Unique race identifier (numeric).}
  \item{state}{Name of the state (character).}
  \item{abbr}{State abbreviation (character).}
  \item{num}{District number (numeric).}
  \item{name1}{Name of the first candidate (character).}
  \item{perc1}{Percentage of votes for the first candidate (numeric).}
  \item{party1}{Party affiliation of the first candidate (character).}
  \item{votes1}{Number of votes for the first candidate (numeric).}
  \item{name2}{Name of the second candidate (character).}
  \item{perc2}{Percentage of votes for the second candidate (numeric).}
  \item{party2}{Party affiliation of the second candidate (character).}
  \item{votes2}{Number of votes for the second candidate (numeric).}
  \item{name3}{Name of the third candidate (character).}
  \item{perc3}{Percentage of votes for the third candidate (numeric).}
  \item{party3}{Party affiliation of the third candidate (character).}
  \item{votes3}{Number of votes for the third candidate (numeric).}
  \item{name4}{Name of the fourth candidate (character).}
  \item{perc4}{Percentage of votes for the fourth candidate (numeric).}
  \item{party4}{Party affiliation of the fourth candidate (character).}
  \item{votes4}{Number of votes for the fourth candidate (numeric).}
  \item{name5}{Name of the fifth candidate (character).}
  \item{perc5}{Percentage of votes for the fifth candidate (numeric).}
  \item{party5}{Party affiliation of the fifth candidate (character).}
  \item{votes5}{Number of votes for the fifth candidate (numeric).}
}
}
\source{
2010 U.S. House of Representatives Election Data
}
\usage{
data(houserace10_tbl_df)
}
\description{
The dataset name has been changed to 'houserace10_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
