% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{immigration_tbl_df}
\alias{immigration_tbl_df}
\title{Poll on illegal workers in the US}
\format{
A tibble with 910 observations and 2 variables:
\describe{
  \item{response}{Factor indicating the response to immigration-related questions, with 4 levels.}
  \item{political}{Factor indicating the political alignment associated with the responses, with 3 levels.}
}
}
\source{
Data from surveys on immigration attitudes
}
\usage{
data(immigration_tbl_df)
}
\description{
The dataset name has been changed to 'immigration_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
