% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{leg_mari_tbl_df}
\alias{leg_mari_tbl_df}
\title{Legalization of Marijuana Support in 2010 California Survey}
\format{
A tibble with 119 observations and 1 variable:
\describe{
  \item{response}{Factor indicating responses related to legal marijuana, with 2 levels.}
}
}
\source{
Data from surveys on attitudes towards legal marijuana
}
\usage{
data(leg_mari_tbl_df)
}
\description{
The dataset name has been changed to 'leg_mari_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
