% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{marathon_tbl_df}
\alias{marathon_tbl_df}
\title{New York City Marathon Times (outdated)}
\format{
A tibble with 59 observations and 3 variables:
\describe{
  \item{year}{Integer indicating the year of the marathon event.}
  \item{gender}{Factor indicating the gender of the participants, with 2 levels.}
  \item{time}{Numeric value representing the marathon completion time in hours.}
}
}
\source{
Data from marathon event results
}
\usage{
data(marathon_tbl_df)
}
\description{
The dataset name has been changed to 'marathon_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
