% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{military_tbl_df}
\alias{military_tbl_df}
\title{US Military Demographics}
\format{
A tibble with an unspecified number of observations and 6 variables:
\describe{
  \item{grade}{Factor indicating the military grade, with 3 levels.}
  \item{branch}{Factor indicating the branch of the military, with 4 levels.}
  \item{gender}{Factor indicating the gender of the participants, with 2 levels.}
  \item{race}{Factor indicating the race of the participants, with 7 levels.}
  \item{hisp}{Logical indicating whether the participants identify as Hispanic.}
  \item{rank}{Integer representing the rank of the participants.}
}
}
\source{
Data from military personnel demographics
}
\usage{
data(military_tbl_df)
}
\description{
The dataset name has been changed to 'military_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
