% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{minn38_df}
\alias{minn38_df}
\title{Minnesota High School Graduates of 1938}
\format{
A data frame with 168 observations and 5 variables:
\describe{
  \item{hs}{Factor indicating the high school status, with 3 levels.}
  \item{phs}{Factor indicating the post-high school status, with 4 levels.}
  \item{fol}{Factor indicating the field of study, with 7 levels.}
  \item{sex}{Factor indicating the gender of the participants, with 2 levels.}
  \item{f}{Integer representing the associated numerical value for the participants.}
}
}
\source{
Data from the Minnesota 1938 study
}
\usage{
data(minn38_df)
}
\description{
The dataset name has been changed to 'minn38_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a data frame. The original content of the dataset has not been modified.
}
