% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{mn_police_use_of_force_df}
\alias{mn_police_use_of_force_df}
\title{Minneapolis police use of force data.}
\format{
A data frame with 12925 observations and 13 variables:
\describe{
  \item{response_datetime}{Character string representing the date and time of the response.}
  \item{problem}{Character string describing the nature of the problem.}
  \item{is_911_call}{Character string indicating whether the incident was initiated by a 911 call.}
  \item{primary_offense}{Character string indicating the primary offense involved in the incident.}
  \item{subject_injury}{Character string describing the injuries sustained by the subject, if any.}
  \item{force_type}{Character string describing the type of force used by the police.}
  \item{force_type_action}{Character string describing the specific actions related to the use of force.}
  \item{race}{Character string indicating the race of the subject involved in the incident.}
  \item{sex}{Character string indicating the sex of the subject.}
  \item{age}{Integer representing the age of the subject.}
  \item{type_resistance}{Character string describing the type of resistance offered by the subject.}
  \item{precinct}{Character string indicating the precinct in which the incident occurred.}
  \item{neighborhood}{Character string representing the neighborhood where the incident occurred.}
}
}
\source{
Data from police use of force reports in Minnesota
}
\usage{
data(mn_police_use_of_force_df)
}
\description{
The dataset name has been changed to 'mn_police_use_of_force_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a data frame. The original content of the dataset has not been modified.
}
