% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{nba_players_19_tbl_df}
\alias{nba_players_19_tbl_df}
\title{NBA Players for the 2018-2019 season}
\format{
A tibble with 494 observations and 7 variables:
\describe{
  \item{first_name}{Character string representing the player's first name.}
  \item{last_name}{Character string representing the player's last name.}
  \item{team}{Character string indicating the name of the team.}
  \item{team_abbr}{Character string representing the team's abbreviation.}
  \item{position}{Character string indicating the player's position on the team.}
  \item{number}{Character string representing the player's jersey number.}
  \item{height}{Numeric value representing the player's height.}
}
}
\source{
Data from NBA players' statistics in 2019
}
\usage{
data(nba_players_19_tbl_df)
}
\description{
The dataset name has been changed to 'nba_players_19_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
