% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{nyc_marathon_tbl_df}
\alias{nyc_marathon_tbl_df}
\title{New York City Marathon Times}
\format{
A tibble with 102 observations and 7 variables:
\describe{
  \item{year}{Numeric value representing the year the marathon took place.}
  \item{name}{Character value representing the name of the runner.}
  \item{country}{Character value indicating the country of origin of the runner.}
  \item{time}{Time variable in 'hms' format representing the finish time of the runner.}
  \item{time_hrs}{Numeric value representing the finish time of the runner in hours.}
  \item{division}{Character value indicating the division (category) the runner participated in.}
  \item{note}{Character value containing additional notes, if any, about the runner or the race.}
}
}
\source{
Data from the New York City Marathon records
}
\usage{
data(nyc_marathon_tbl_df)
}
\description{
The dataset name has been changed to 'nyc_marathon_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
