% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{nycvehiclethefts_tbl_df}
\alias{nycvehiclethefts_tbl_df}
\title{Thefts of motor vehicles 2014 to 2017}
\format{
A tibble with 35,746 observations and 9 variables:
\describe{
  \item{uid}{Integer value representing a unique identifier for each vehicle theft incident.}
  \item{date_single}{Character value representing the single date of the theft incident.}
  \item{date_start}{Character value representing the start date of the theft incident.}
  \item{date_end}{Character value representing the end date of the theft incident.}
  \item{longitude}{Numeric value indicating the longitude where the incident occurred.}
  \item{latitude}{Numeric value indicating the latitude where the incident occurred.}
  \item{location_type}{Character value representing the type of location where the theft took place.}
  \item{location_category}{Character value indicating the category of the location.}
  \item{census_block}{Character value indicating the census block where the incident took place.}
}
}
\source{
Data from the New York City Vehicle Thefts records
}
\usage{
data(nycvehiclethefts_tbl_df)
}
\description{
The dataset name has been changed to 'nycvehiclethefts_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
