% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{offshore_drilling_tbl_df}
\alias{offshore_drilling_tbl_df}
\title{California poll on drilling off the California coast}
\format{
A tibble with 828 observations and 2 variables:
\describe{
  \item{v1}{Factor with 4 levels, representing different responses or categories related to offshore drilling.}
  \item{v2}{Factor with 3 levels, representing secondary categories or classifications related to the responses in \code{v1}.}
}
}
\source{
Data related to offshore drilling opinions or classifications
}
\usage{
data(offshore_drilling_tbl_df)
}
\description{
The dataset name has been changed to 'offshore_drilling_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
