% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{orings_tbl_df}
\alias{orings_tbl_df}
\title{1986 Challenger disaster and O-rings}
\format{
A tibble with 23 observations and 4 variables:
\describe{
  \item{mission}{Integer representing the mission number.}
  \item{temperature}{Integer representing the launch temperature in Fahrenheit.}
  \item{damaged}{Integer representing the number of damaged O-rings in the mission.}
  \item{undamaged}{Numeric representing the number of undamaged O-rings in the mission.}
}
}
\source{
Data from NASA missions related to O-ring performance.
}
\usage{
data(orings_tbl_df)
}
\description{
The dataset name has been changed to 'orings_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
