% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{oscars_tbl_df}
\alias{oscars_tbl_df}
\title{Oscar winners, 1929 to 2018}
\format{
A tibble with 184 observations and 11 variables:
\describe{
  \item{oscar_no}{Numeric indicating the Oscar number.}
  \item{oscar_yr}{Numeric representing the year the Oscar was awarded.}
  \item{award}{Character string indicating the category of the award.}
  \item{name}{Character string with the name of the recipient.}
  \item{movie}{Character string indicating the movie for which the award was given.}
  \item{age}{Numeric indicating the age of the recipient at the time of the award.}
  \item{birth_pl}{Character string indicating the birthplace of the recipient.}
  \item{birth_date}{Date representing the birthdate of the recipient.}
  \item{birth_mo}{Numeric indicating the birth month.}
  \item{birth_d}{Numeric indicating the birth day.}
  \item{birth_y}{Numeric indicating the birth year.}
}
}
\source{
Data from historical Oscar award records.
}
\usage{
data(oscars_tbl_df)
}
\description{
The dataset name has been changed to 'oscars_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
