% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{piracy_tbl_df}
\alias{piracy_tbl_df}
\title{Piracy and PIPA/SOPA}
\format{
A tibble with 534 observations and 8 variables:
\describe{
  \item{name}{Character string indicating the name of the politician.}
  \item{party}{Factor with 3 levels representing the politician's party affiliation.}
  \item{state}{Factor with 50 levels indicating the U.S. state the politician represents.}
  \item{money_pro}{Numeric representing the amount of pro-piracy funding received.}
  \item{money_con}{Numeric representing the amount of anti-piracy funding received.}
  \item{years}{Integer indicating the number of years in office.}
  \item{stance}{Factor with 5 levels indicating the politician's stance on piracy.}
  \item{chamber}{Factor with 2 levels indicating the chamber of the U.S. Congress (House or Senate).}
}
}
\source{
Data on political stances and funding related to piracy.
}
\usage{
data(piracy_tbl_df)
}
\description{
The dataset name has been changed to 'piracy_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
