% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{precip_numeric}
\alias{precip_numeric}
\title{Annual Precipitation in US Cities}
\format{
A numeric vector with 70 observations representing average annual precipitation (in inches)
for various cities in the United States.
\describe{
  \item{Mobile}{Numeric value representing the average annual precipitation in Mobile.}
  \item{Juneau}{Numeric value representing the average annual precipitation in Juneau.}
  \item{Phoenix}{Numeric value representing the average annual precipitation in Phoenix.}
  \item{Los Angeles}{Numeric value representing the average annual precipitation in Los Angeles.}
  \item{...}{Additional cities included in the dataset.}
}
}
\source{
Data on precipitation for various U.S. cities.
}
\usage{
data(precip_numeric)
}
\description{
The dataset name has been changed to 'precip_numeric' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a numeric vector. The original content of the dataset has not been modified.
}
