% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{presidents_ts}
\alias{presidents_ts}
\title{Quarterly Approval Ratings of US Presidents}
\format{
A time series object with 120 observations, covering quarterly data from 1945 to 1975.
Each observation represents the number of presidents' approval ratings during a given quarter.
The data is structured as follows:
\describe{
  \item{Qtr1}{Numeric values representing the approval ratings for the first quarter.}
  \item{Qtr2}{Numeric values representing the approval ratings for the second quarter.}
  \item{Qtr3}{Numeric values representing the approval ratings for the third quarter.}
  \item{Qtr4}{Numeric values representing the approval ratings for the fourth quarter.}
}
}
\source{
Data on presidential approval ratings from 1945 to 1975.
}
\usage{
data(presidents_ts)
}
\description{
The dataset name has been changed to 'presidents_ts' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a time series object. The original content of the dataset has not been modified.
}
