% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{prrace08_tbl_df}
\alias{prrace08_tbl_df}
\title{Election results for the 2008 U.S. Presidential race}
\format{
A tibble with 51 observations and 7 variables:
\describe{
  \item{state}{Factor indicating the U.S. state (including Washington D.C.) where the election took place.}
  \item{state_full}{Factor providing the full name of the U.S. state corresponding to the abbreviation.}
  \item{n_obama}{Integer representing the number of votes received by Barack Obama in the state.}
  \item{p_obama}{Numeric representing the percentage of total votes received by Barack Obama in the state.}
  \item{n_mc_cain}{Integer representing the number of votes received by John McCain in the state.}
  \item{p_mc_cain}{Numeric representing the percentage of total votes received by John McCain in the state.}
  \item{el_votes}{Integer indicating the number of electoral votes allocated to the state.}
}
}
\source{
Data on the 2008 U.S. presidential race results by state.
}
\usage{
data(prrace08_tbl_df)
}
\description{
The dataset name has been changed to 'prrace08_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
