% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{road_df}
\alias{road_df}
\title{Road Accident Deaths in US States}
\format{
A data frame with 26 observations and 6 variables:
\describe{
  \item{deaths}{Integer indicating the number of road deaths.}
  \item{drivers}{Integer representing the number of licensed drivers.}
  \item{popden}{Numeric indicating the population density (people per square mile).}
  \item{rural}{Numeric indicating the percentage of rural roads.}
  \item{temp}{Integer representing the average temperature (in degrees Fahrenheit).}
  \item{fuel}{Numeric indicating the fuel consumption per capita (in gallons).}
}
}
\source{
Data on road safety statistics, including deaths, drivers, population density, and environmental factors.
}
\usage{
data(road_df)
}
\description{
The dataset name has been changed to 'road_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a data frame. The original content of the dataset has not been modified.
}
