% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{senaterace10_tbl_df}
\alias{senaterace10_tbl_df}
\title{Election results for the 2010 U.S. Senate races}
\format{
A tibble with 38 observations and 23 variables:
\describe{
  \item{id}{Numeric identifier for the election race.}
  \item{state}{Character string indicating the U.S. state where the election took place.}
  \item{abbr}{Character string representing the state abbreviation.}
  \item{name1}{Character string indicating the name of the first candidate.}
  \item{perc1}{Numeric indicating the percentage of votes received by the first candidate.}
  \item{party1}{Character string indicating the party affiliation of the first candidate.}
  \item{votes1}{Numeric indicating the total votes received by the first candidate.}
  \item{name2}{Character string indicating the name of the second candidate.}
  \item{perc2}{Numeric indicating the percentage of votes received by the second candidate.}
  \item{party2}{Character string indicating the party affiliation of the second candidate.}
  \item{votes2}{Numeric indicating the total votes received by the second candidate.}
  \item{name3}{Character string indicating the name of the third candidate.}
  \item{perc3}{Numeric indicating the percentage of votes received by the third candidate.}
  \item{party3}{Character string indicating the party affiliation of the third candidate.}
  \item{votes3}{Numeric indicating the total votes received by the third candidate.}
  \item{name4}{Character string indicating the name of the fourth candidate.}
  \item{perc4}{Numeric indicating the percentage of votes received by the fourth candidate.}
  \item{party4}{Character string indicating the party affiliation of the fourth candidate.}
  \item{votes4}{Numeric indicating the total votes received by the fourth candidate.}
  \item{name5}{Character string indicating the name of the fifth candidate.}
  \item{perc5}{Numeric indicating the percentage of votes received by the fifth candidate.}
  \item{party5}{Character string indicating the party affiliation of the fifth candidate.}
  \item{votes5}{Numeric indicating the total votes received by the fifth candidate.}
}
}
\source{
Data on U.S. Senate races held in 2010, including candidates' names, vote percentages, and party affiliations.
}
\usage{
data(senaterace10_tbl_df)
}
\description{
The dataset name has been changed to 'senaterace10_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
