% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{sp500_1950_2018_tbl_df}
\alias{sp500_1950_2018_tbl_df}
\title{Daily observations for the S&P 500 - Historical Data (1950-2018)}
\format{
A tibble with 17346 observations and 7 variables:
\describe{
  \item{Date}{Factor indicating the date of the recorded stock prices.}
  \item{Open}{Numeric representing the opening price of the stock.}
  \item{High}{Numeric representing the highest price of the stock during the day.}
  \item{Low}{Numeric representing the lowest price of the stock during the day.}
  \item{Close}{Numeric representing the closing price of the stock.}
  \item{Adj.Close}{Numeric representing the adjusted closing price of the stock.}
  \item{Volume}{Numeric representing the trading volume of the stock.}
}
}
\source{
Historical data on S&P 500 stock prices from 1950 to 2018.
}
\usage{
data(sp500_1950_2018_tbl_df)
}
\description{
The dataset name has been changed to 'sp500_1950_2018_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
