% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{state_area_numeric}
\alias{state_area_numeric}
\title{US State Facts and Figures - US State Areas}
\format{
A numeric dataset with 50 elements representing the area of U.S. states in square kilometers:
\describe{
  \item{state_area}{Numeric values indicating the area of each state, measured in square kilometers.}
}
}
\source{
U.S. state areas.
}
\usage{
data(state_area_numeric)
}
\description{
The dataset name has been changed to 'state_area_numeric' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a numeric dataset. The original content of the dataset has not been modified.
}
