% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{state_center_list}
\alias{state_center_list}
\title{US State Facts and Figures - US State Centers}
\format{
A list with 2 elements, each containing numeric values representing the geographical center coordinates
of U.S. states:
\describe{
  \item{x}{Numeric vector of length 50 representing the x-coordinates (longitude) of the state centers.}
  \item{y}{Numeric vector of length 50 representing the y-coordinates (latitude) of the state centers.}
}
}
\source{
Geographical data for U.S. state centers.
}
\usage{
data(state_center_list)
}
\description{
The dataset name has been changed to 'state_center_list' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a list. The original content of the dataset has not been modified.
}
