% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{state_name_character}
\alias{state_name_character}
\title{US State Facts and Figures - US State Names}
\format{
A character vector with 50 observations representing the names of U.S. states.
\describe{
  \item{"Alabama"}{Name of the state.}
  \item{"Alaska"}{Name of the state.}
  \item{"Arizona"}{Name of the state.}
  \item{"Arkansas"}{Name of the state.}
  \item{"California"}{Name of the state.}
  \item{"Colorado"}{Name of the state.}
  \item{"Connecticut"}{Name of the state.}
  \item{"Delaware"}{Name of the state.}
  \item{"Florida"}{Name of the state.}
  \item{"Georgia"}{Name of the state.}
  \item{"Hawaii"}{Name of the state.}
  \item{"Idaho"}{Name of the state.}
  \item{"Illinois"}{Name of the state.}
  \item{"Indiana"}{Name of the state.}
  \item{"Iowa"}{Name of the state.}
  \item{"Kansas"}{Name of the state.}
  \item{"Kentucky"}{Name of the state.}
  \item{"Louisiana"}{Name of the state.}
  \item{"Maine"}{Name of the state.}
  \item{"Maryland"}{Name of the state.}
  \item{"Massachusetts"}{Name of the state.}
  \item{"Michigan"}{Name of the state.}
  \item{"Minnesota"}{Name of the state.}
  \item{"Mississippi"}{Name of the state.}
  \item{"Missouri"}{Name of the state.}
  \item{"Montana"}{Name of the state.}
  \item{"Nebraska"}{Name of the state.}
  \item{"Nevada"}{Name of the state.}
  \item{"New Hampshire"}{Name of the state.}
  \item{"New Jersey"}{Name of the state.}
  \item{"New Mexico"}{Name of the state.}
  \item{"New York"}{Name of the state.}
  \item{"North Carolina"}{Name of the state.}
  \item{"North Dakota"}{Name of the state.}
  \item{"Ohio"}{Name of the state.}
  \item{"Oklahoma"}{Name of the state.}
  \item{"Oregon"}{Name of the state.}
  \item{"Pennsylvania"}{Name of the state.}
  \item{"Rhode Island"}{Name of the state.}
  \item{"South Carolina"}{Name of the state.}
  \item{"South Dakota"}{Name of the state.}
  \item{"Tennessee"}{Name of the state.}
  \item{"Texas"}{Name of the state.}
  \item{"Utah"}{Name of the state.}
  \item{"Vermont"}{Name of the state.}
  \item{"Virginia"}{Name of the state.}
  \item{"Washington"}{Name of the state.}
  \item{"West Virginia"}{Name of the state.}
  \item{"Wisconsin"}{Name of the state.}
  \item{"Wyoming"}{Name of the state.}
}
}
\source{
U.S. Census Bureau.
}
\usage{
data(state_name_character)
}
\description{
The dataset name has been changed to 'state_name_character' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a character vector. The original content of the dataset has not been modified.
}
