% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{state_region_factor}
\alias{state_region_factor}
\title{US State Facts and Figures - US State Regions}
\format{
A factor variable with 50 observations, representing the region of each U.S. state.
The regions are classified into four levels:
\describe{
  \item{"Northeast"}{States located in the Northeast region.}
  \item{"South"}{States located in the Southern region.}
  \item{"North Central"}{States located in the North Central region.}
  \item{"West"}{States located in the Western region.}
}
}
\source{
U.S. Census Bureau.
}
\usage{
data(state_region_factor)
}
\description{
The dataset name has been changed to 'state_region_factor' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a factor variable representing U.S. state regions.
}
