% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{state_x77_matrix}
\alias{state_x77_matrix}
\title{US State Facts and Figures - US State Demographics and Statistics (1977)}
\format{
A matrix with 50 rows and 8 columns representing various demographic and statistical
characteristics of U.S. states. The columns include:
\describe{
  \item{Population}{Population of the state.}
  \item{Income}{Median income of the state's residents.}
  \item{Illiteracy}{Illiteracy rate (percentage).}
  \item{Life Exp}{Life expectancy (in years).}
  \item{Murder}{Murder rate (per 100,000 inhabitants).}
  \item{HS Grad}{High school graduation rate (percentage).}
  \item{Frost}{Number of days with frost.}
  \item{Area}{Total area of the state (in square miles).}
}
}
\source{
U.S. Census Bureau (1977).
}
\usage{
data(state_x77_matrix)
}
\description{
The dataset name has been changed to 'state_x77_matrix' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a matrix variable representing various demographic and statistical
attributes of U.S. states in 1977.
}
