% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{UCBAdmissions_table}
\alias{UCBAdmissions_table}
\title{Student Admissions at UC Berkeley}
\format{
A table object with 24 entries representing the admissions data at U.C. Berkeley:
\describe{
  \item{Admit}{A factor with levels "Admitted" and "Rejected".}
  \item{Gender}{A factor with levels "Male" and "Female".}
  \item{Dept}{A factor representing the department with levels "A", "B", "C", "D", "E", and "F".}
  \item{values}{Numeric counts of admissions based on gender and department.}
}
}
\source{
U.C. Berkeley admissions data from 1973.
}
\usage{
data(UCBAdmissions_table)
}
\description{
The dataset name has been changed to 'UCBAdmissions_table' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a table object. The original content of the dataset has not been modified.
}
