% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{us_crime_rates_spec_tbl_df}
\alias{us_crime_rates_spec_tbl_df}
\title{US Crime Rates}
\format{
A tibble with 60 rows and 12 columns:
\describe{
  \item{year}{Numeric year of the recorded data, e.g., 2000, 2001.}
  \item{population}{Numeric population total for the respective year.}
  \item{total}{Numeric total number of crimes reported.}
  \item{violent}{Numeric total number of violent crimes.}
  \item{property}{Numeric total number of property crimes.}
  \item{murder}{Numeric total number of murders.}
  \item{forcible_rape}{Numeric total number of forcible rapes.}
  \item{robbery}{Numeric total number of robberies.}
  \item{aggravated_assault}{Numeric total number of aggravated assaults.}
  \item{burglary}{Numeric total number of burglaries.}
  \item{larceny_theft}{Numeric total number of larcenies.}
  \item{vehicle_theft}{Numeric total number of vehicle thefts.}
}
}
\source{
Federal Bureau of Investigation (FBI) Uniform Crime Reporting (UCR) Program.
}
\usage{
data(us_crime_rates_spec_tbl_df)
}
\description{
The dataset 'us_crime_rates_spec_tbl_df' contains crime statistics for the United States,
including various types of crimes and population data for each year. This dataset is
structured as a tibble for ease of use within the R ecosystem. This naming convention helps
distinguish this dataset as part of the 'usdatasets' package.
}
