% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{us_time_survey_tbl_df}
\alias{us_time_survey_tbl_df}
\title{American Time Survey 2009 - 2019}
\format{
A tibble with 11 observations and 8 variables representing time use in various activities:
\describe{
  \item{year}{Numeric value representing the year of the survey.}
  \item{household_activities}{Numeric value representing time spent on household activities (in hours).}
  \item{eating_and_drinking}{Numeric value representing time spent on eating and drinking (in hours).}
  \item{leisure_and_sports}{Numeric value representing time spent on leisure and sports activities (in hours).}
  \item{sleeping}{Numeric value representing time spent sleeping (in hours).}
  \item{caring_children}{Numeric value representing time spent caring for children (in hours).}
  \item{working_employed}{Numeric value representing time spent working while employed (in hours).}
  \item{working_employed_days_worked}{Numeric value representing the number of days worked while employed.}
}
}
\source{
U.S. Bureau of Labor Statistics.
}
\usage{
data(us_time_survey_tbl_df)
}
\description{
The dataset name has been changed to 'us_time_survey_tbl_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a tibble. The original content of the dataset has not been modified.
}
