% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USAccDeaths_ts}
\alias{USAccDeaths_ts}
\title{Accidental Deaths in the US 1973-1978}
\format{
A time series object with 72 observations representing monthly accidental deaths in the U.S.
from 1973 to 1979:
\describe{
  \item{years}{A numeric vector representing the years from 1973 to 1979.}
  \item{months}{A character vector representing the months from January to December.}
  \item{deaths}{Numeric values representing the number of accidental deaths for each month.}
}
}
\source{
U.S. accidental deaths data.
}
\usage{
data(USAccDeaths_ts)
}
\description{
The dataset name has been changed to 'USAccDeaths_ts' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a time series object. The original content of the dataset has not been modified.
}
