% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USArrests_df}
\alias{USArrests_df}
\title{Violent Crime Rates by US State}
\format{
A data frame with 50 observations and 4 variables representing the rates of arrests in the U.S.:
\describe{
  \item{Murder}{Numeric vector representing the murder rates per 100,000 residents.}
  \item{Assault}{Integer vector representing the assault rates per 100,000 residents.}
  \item{UrbanPop}{Integer vector representing the percentage of the population living in urban areas.}
  \item{Rape}{Numeric vector representing the rape rates per 100,000 residents.}
}
}
\source{
U.S. arrests data from 1973.
}
\usage{
data(USArrests_df)
}
\description{
The dataset name has been changed to 'USArrests_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a data frame. The original content of the dataset has not been modified.
}
