% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{UScitiesD_dist}
\alias{UScitiesD_dist}
\title{Distances Between European Cities and Between US Cities}
\format{
A distance object containing the distances (in miles) between selected U.S. cities:
\describe{
  \item{Atlanta}{Distance from Atlanta to other cities.}
  \item{Chicago}{Distance from Chicago to other cities.}
  \item{Denver}{Distance from Denver to other cities.}
  \item{Houston}{Distance from Houston to other cities.}
  \item{LosAngeles}{Distance from Los Angeles to other cities.}
  \item{Miami}{Distance from Miami to other cities.}
  \item{NewYork}{Distance from New York to other cities.}
  \item{SanFrancisco}{Distance from San Francisco to other cities.}
  \item{Seattle}{Distance from Seattle to other cities.}
  \item{Washington.DC}{Distance from Washington D.C. to other cities.}
}
}
\source{
U.S. cities distance data.
}
\usage{
data(UScitiesD_dist)
}
\description{
The dataset name has been changed to 'UScitiesD_dist' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a distance object. The original content of the dataset has not been modified.
}
