% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdatasets-package.R
\docType{package}
\name{usdatasets}
\alias{usdatasets}
\alias{usdatasets-package}
\title{usdatasets: A Comprehensive Collection of U.S. Datasets}
\description{
{
This package provides a wide variety of datasets related to crime, economy, society, politics, and sports within the United States for testing, learning, and research purposes.
}
}
\details{
usdatasets: A Comprehensive Collection of U.S. Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of U.S. Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/usdatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Cáceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
