% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USJudgeRatings_df}
\alias{USJudgeRatings_df}
\title{Lawyers' Ratings of State Judges in the US Superior Court}
\format{
A data frame with 43 observations and 12 variables representing ratings for U.S. judges:
\describe{
  \item{CONT}{Numeric vector representing the judges' ratings on control.}
  \item{INTG}{Numeric vector representing the judges' ratings on integrity.}
  \item{DMNR}{Numeric vector representing the judges' ratings on demeanor.}
  \item{DILG}{Numeric vector representing the judges' ratings on diligence.}
  \item{CFMG}{Numeric vector representing the judges' ratings on communications with clients.}
  \item{DECI}{Numeric vector representing the judges' ratings on decisiveness.}
  \item{PREP}{Numeric vector representing the judges' ratings on preparation.}
  \item{FAMI}{Numeric vector representing the judges' ratings on family law expertise.}
  \item{ORAL}{Numeric vector representing the judges' ratings on oral communications.}
  \item{WRIT}{Numeric vector representing the judges' ratings on written communications.}
  \item{PHYS}{Numeric vector representing the judges' ratings on physical appearance.}
  \item{RTEN}{Numeric vector representing the judges' ratings on overall rating.}
}
}
\source{
U.S. judge ratings data.
}
\usage{
data(USJudgeRatings_df)
}
\description{
The dataset name has been changed to 'USJudgeRatings_df' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a data frame. The original content of the dataset has not been modified.
}
