% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USPersonalExpenditure_matrix}
\alias{USPersonalExpenditure_matrix}
\title{Personal Expenditure Data}
\format{
A matrix with 5 rows and 5 columns representing U.S. personal expenditures in different categories over selected years:
\describe{
  \item{Food and Tobacco}{Numeric values representing expenditures on food and tobacco for the years 1940, 1945, 1950, 1955, and 1960.}
  \item{Household Operation}{Numeric values representing expenditures on household operations for the years 1940, 1945, 1950, 1955, and 1960.}
  \item{Medical and Health}{Numeric values representing expenditures on medical and health services for the years 1940, 1945, 1950, 1955, and 1960.}
  \item{Personal Care}{Numeric values representing expenditures on personal care for the years 1940, 1945, 1950, 1955, and 1960.}
  \item{Private Education}{Numeric values representing expenditures on private education for the years 1940, 1945, 1950, 1955, and 1960.}
}
}
\source{
U.S. personal expenditure data.
}
\usage{
data(USPersonalExpenditure_matrix)
}
\description{
The dataset name has been changed to 'USPersonalExpenditure_matrix' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a matrix. The original content of the dataset has not been modified.
}
