% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{uspop_ts}
\alias{uspop_ts}
\title{Populations Recorded by the US Census}
\format{
A time series object with 19 observations representing the U.S. population from 1790 to 1970:
\describe{
  \item{values}{Numeric vector containing the population values in millions.}
}
}
\source{
U.S. Census Bureau.
}
\usage{
data(uspop_ts)
}
\description{
The dataset name has been changed to 'uspop_ts' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a time series object. The original content of the dataset has not been modified.
}
