% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{VADeaths_matrix}
\alias{VADeaths_matrix}
\title{Death Rates in Virginia (1940)}
\format{
A matrix containing mortality rates (per 1000) for different demographic groups in Virginia:
\describe{
  \item{Rural Male}{Mortality rates for rural males by age group.}
  \item{Rural Female}{Mortality rates for rural females by age group.}
  \item{Urban Male}{Mortality rates for urban males by age group.}
  \item{Urban Female}{Mortality rates for urban females by age group.}
}
}
\source{
Virginia mortality data.
}
\usage{
data(VADeaths_matrix)
}
\description{
The dataset name has been changed to 'VADeaths_matrix' to avoid confusion with other packages
in the R ecosystem. This naming convention helps distinguish this dataset as part of the 'usdatasets'
package and identifies it as a matrix. The original content of the dataset has not been modified.
}
